refactor: move binding models to metadata folder
