+++ packages/iocuak/src/container/services/api/ContainerBindingApiService.ts
@@ -1,6 +1,6 @@
-import { BindingApi } from '../../../binding/models/api/BindingApi';
 import { Newable } from '../../../common/models/domain/Newable';
 import { ServiceId } from '../../../common/models/domain/ServiceId';
+import { BindingApi } from '../../../metadata/models/api/BindingApi';
 
 export interface ContainerBindingApiService {
   bind<TInstance, TArgs extends unknown[]>(
