+++ packages/iocuak/src/container/services/domain/ContainerMetadataService.ts
@@ -1,6 +1,6 @@
-import { TypeBinding } from '../../../binding/models/domain/TypeBinding';
 import { Newable } from '../../../common/models/domain/Newable';
 import { ClassMetadata } from '../../../metadata/models/domain/ClassMetadata';
+import { TypeBinding } from '../../../metadata/models/domain/TypeBinding';
 
 export interface ContainerMetadataService {
   getBindingMetadata<TInstance, TArgs extends unknown[]>(
