+++ packages/iocuak/src/index.ts
@@ -1,6 +1,3 @@
-import { BindingApi } from './binding/models/api/BindingApi';
-import { TypeBindingApi } from './binding/models/api/TypeBindingApi';
-import { ValueBindingApi } from './binding/models/api/ValueBindingApi';
 import { Newable } from './common/models/domain/Newable';
 import { ServiceId } from './common/models/domain/ServiceId';
 import { ContainerApi } from './container/modules/api/ContainerApi';
@@ -10,7 +7,10 @@ import { inject } from './metadata/decorators/inject';
 import { injectable } from './metadata/decorators/injectable';
 import { injectFrom } from './metadata/decorators/injectFrom';
 import { injectFromBase } from './metadata/decorators/injectFromBase';
+import { BindingApi } from './metadata/models/api/BindingApi';
 import { InjectableOptionsApi } from './metadata/models/api/InjectableOptionsApi';
+import { TypeBindingApi } from './metadata/models/api/TypeBindingApi';
+import { ValueBindingApi } from './metadata/models/api/ValueBindingApi';
 import { TaskScopeApi } from './task/models/api/TaskScopeApi';
 
 export type {
