+++ packages/iocuak/src/metadata/decorators/injectFromBase.spec.ts
@@ -1,9 +1,9 @@
 jest.mock('../../utils/getBaseType');
 jest.mock('./injectFrom');
 
-import { ClassMetadataExtensionApi } from '../../binding/models/api/ClassMetadataExtensionApi';
 import { Newable } from '../../common/models/domain/Newable';
 import { getBaseType } from '../../utils/getBaseType';
+import { ClassMetadataExtensionApi } from '../models/api/ClassMetadataExtensionApi';
 import { injectFrom } from './injectFrom';
 import { injectFromBase } from './injectFromBase';
